backend/requirements.txt
# FastAPI and ASGI server
fastapi==0.104.1
uvicorn[standard]==0.24.0
gunicorn==21.2.0

# Database and ORM
supabase==2.0.2
psycopg2-binary==2.9.9

# Authentication and Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6

# Environment and Configuration
python-dotenv==1.0.0
pydantic==2.5.0
pydantic-settings==2.1.0

# HTTP and CORS
httpx>=0.24.0,<0.25.0

# Logging and Monitoring
structlog==23.2.0
sentry-sdk[fastapi]==1.38.0

# Performance and Caching
redis==5.0.1
aioredis==2.0.1

# Development and Testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0

# Utilities
python-dateutil==2.8.2
pytz==2023.3
email-validator==2.1.0

# Production Dependencies
whitenoise==6.6.0

# Optional: For advanced features
celery==5.3.4
flower==2.0.1
